% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressify.R
\name{progressify}
\alias{progressify}
\alias{pz}
\title{Evaluate a regular map-reduce call with progress updates}
\usage{
progressify(
  expr,
  substitute = TRUE,
  ...,
  when = TRUE,
  eval = TRUE,
  envir = parent.frame()
)
}
\arguments{
\item{expr}{An R expression.}

\item{substitute}{If TRUE, \code{expr} is quoted.}

\item{when}{If TRUE (default), the expression is progressified, otherwise not.}

\item{eval}{If TRUE (default), the progressified expression is evaluated,
otherwise it is returned.}

\item{envir}{The environment in which \code{expr} is evaluated.}

\item{\ldots}{Not used.}
}
\value{
Returns the value of the evaluated expression \code{expr}.
}
\description{
Evaluate a regular map-reduce call with progress updates
}
\section{Expression unwrapping}{

The transpilation mechanism includes logic to "unwrap" expressions
enclosed in constructs such as \code{!}, \code{{ }}, \verb{( )}, \code{local()}, \code{I()},
\code{identity()}, \code{invisible()}, \code{suppressMessages()}, \code{suppressWarnings()},
\code{suppressPackageStartupMessages()}, \code{withCallingHandlers()}, and
\code{with()}. The transpiler descends through wrapping
constructs until it finds a transpilable expression, avoiding the
need to place \code{progressify()} inside such constructs. This allows for
patterns like:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{y <- \{
  lapply(xs, fcn)
\} |> suppressMessages() |> progressify()
}\if{html}{\out{</div>}}

avoiding having to write:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{y <- \{
  lapply(xs, fcn) |> progressify()
\} |> suppressMessages()
}\if{html}{\out{</div>}}
}

\examples{
handlers(global = TRUE) # listen to progress updates

xs <- list(1, 1:2, 1:2, 1:5)
y <- lapply(X = xs, FUN = sum) |> progressify()
str(y)
}
