% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressify_supported_tools.R
\name{progressify_supported_packages}
\alias{progressify_supported_packages}
\alias{progressify_supported_functions}
\title{List packages and functions supporting progressification}
\usage{
progressify_supported_packages()

progressify_supported_functions(package)
}
\arguments{
\item{package}{A package name.}
}
\value{
A character vector of package or function names.
}
\description{
List packages and functions supporting progressification
}
\examples{
pkgs <- progressify_supported_packages()
pkgs

fcns <- progressify_supported_functions("base")
print(fcns)

if (requireNamespace("purrr")) {
  fcns <- progressify_supported_functions("purrr")
  print(fcns)
}

}
