% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressify.R
\name{progressify}
\alias{progressify}
\title{Evaluate a regular map-reduce call in parallel}
\usage{
progressify(expr, substitute = TRUE, ..., envir = parent.frame())
}
\arguments{
\item{expr}{An R expression.}

\item{substitute}{If TRUE, \code{expr} is quoted.}

\item{envir}{The environment in which \code{expr} is evaluated.}

\item{\ldots}{Not used.}
}
\value{
Returns the value of the evaluated expression \code{expr}.
}
\description{
Evaluate a regular map-reduce call in parallel
}
\examples{
xs <- list(1, 1:2, 1:2, 1:5)

# ------------------------------------------
# Base R apply functions
# ------------------------------------------
y <- lapply(X = xs, FUN = function(x) {
  sum(x)
}) |> progressify()
str(y)
}
